Rename stake task to initialize
The task does more than just a stake. It is used to inialize staking for
the Random Beacon for operator and other roles.
