+++ solidity/random-beacon/tasks/index.ts
@@ -1,4 +1,4 @@
 // eslint-disable-next-line global-require
 require("./unlock-eth-accounts")
-require("./stake")
+require("./initialize")
 require("./genesis")
